requirements.txt
blinker==1.7.0
bottle==0.12.25
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
Flask-Cors==4.0.0
Flask-WTF==1.2.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
paho-mqtt==2.0.0
proxy_tools==0.1.0
pycairo==1.26.0
PyGObject==3.48.2
pypng==0.20220715.0
python-dateutil==2.9.0.post0
pywebview==5.0.5
qrcode==7.4.2
requests==2.31.0
screeninfo==0.8.1
six==1.16.0
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.2
WTForms==3.1.2
